peer review
Very interesting to look at this dataset! Your leaflet map should work if you change your noInsp_coords variable to call for the 15th and 14th columns. I did that and knit it and the cluster map worked perfectly. It would be interesting to see if there is anything that can be taken from the no inspection data for parking tickets that excludes the CBD aand areas of high traffic, only looking at residential neighborhoods to see if there is are any interesting correlations.

Best,
Nino